pyproject.toml
[tool.poetry]
name = "image-analysis-mcp"
version = "1.0.0"
description = "MCP server for comprehensive read-only image analysis"
authors = ["Jim Dasher"]
readme = "README.md"
packages = [{include = "image_analysis_mcp", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastmcp = "^1.0.0"
numpy = "^1.26.0"
opencv-python = "^4.9.0"
pillow = "^10.2.0"
rawpy = "^0.19.0"
scikit-image = "^0.22.0"
scikit-learn = "^1.3.0"
exifread = "^3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
image-analysis-mcp = "image_analysis_mcp.server:main"
